+++ duckscript_sdk/src/sdk/std/fs/get_last_modified_time/mod.rs
@@ -1,0 +1,46 @@
+use crate::utils::{io, pckg};
+use duckscript::types::command::{Command, CommandResult};
+
+#[cfg(test)]
+#[path = "./mod_test.rs"]
+mod mod_test;
+
+#[derive(Clone)]
+pub(crate) struct CommandImpl {
+    package: String,
+}
+
+impl Command for CommandImpl {
+    fn name(&self) -> String {
+        pckg::concat(&self.package, "GetLastModifiedTime")
+    }
+
+    fn aliases(&self) -> Vec<String> {
+        vec!["get_last_modified_time".to_string()]
+    }
+
+    fn help(&self) -> String {
+        include_str!("help.md").to_string()
+    }
+
+    fn clone_and_box(&self) -> Box<dyn Command> {
+        Box::new((*self).clone())
+    }
+
+    fn run(&self, arguments: Vec<String>) -> CommandResult {
+        if arguments.is_empty() {
+            CommandResult::Error("Path not provided.".to_string())
+        } else {
+            match io::get_last_modified_time(&arguments[0]) {
+                Ok(time) => CommandResult::Continue(Some(time.to_string())),
+                Err(error) => CommandResult::Error(error),
+            }
+        }
+    }
+}
+
+pub(crate) fn create(package: &str) -> Box<dyn Command> {
+    Box::new(CommandImpl {
+        package: package.to_string(),
+    })
+}